kinesis-data-analytics/kda-flink/msk-replication/requirements.txt
aws-cdk.aws-ec2==1.98.0
aws-cdk.aws-iam==1.98.0
aws-cdk.aws-kinesisanalytics==1.98.0
aws-cdk.aws-logs==1.98.0
aws-cdk.aws-msk==1.98.0
aws-cdk.aws-s3==1.98.0
aws-cdk.core==1.98.0
